+++ RELEASES.md
@@ -13,6 +13,7 @@ Enhancements
 ------------
 
 * Source information is provided when there is an error within a call to a function defined in another module.
+* Source information is provided to the `NativeCallContext` for native Rust functions.
 
 
 Version 0.19.9